+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Changelog.rst
@@ -15,6 +15,7 @@ Fixed
 *****
 - Fix code generator configuration to search GameFinder scene with the proper service type (gamefinder and not gamesession).
 - Add AreCompatible event handlers
+- Fix AreCompatible event handlers
 
 7.0.0
 -----